Project/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.2.1
backcall==0.2.0
blinker==1.7.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distro==1.8.0
executing==1.2.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.39.4
funcy==2.0
gensim==4.3.1
gevent==23.9.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
ipykernel==6.23.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mysql-connector-python==8.2.0
nest-asyncio==1.5.6
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.2
openai==1.2.1
packaging==23.1
pandas==2.0.2
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.15.1
pyLDAvis==3.4.1
pylightxl==1.61
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
scikit-learn==1.2.2
scikit-surprise==1.1.3
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.2
tabulate==0.9.0
threadpoolctl==3.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
websocket==0.2.1
Werkzeug==3.0.1
wordcloud==1.9.2
zope.event==5.0
zope.interface==6.1
